requirements.txt
SQLAlchemy>=2.0,<3.0
pymysql>=1.1.0
python-dotenv>=1.0.0
